Fix ubsan warnings: vp9/encoder/vp9_mcomp.c
This commit fixes a number of ubsan warnings in HBD build.

BUG=webm:1219

+++ vp9/decoder/vp9_decodeframe.c
@@ -1651,7 +1651,7 @@ static int tile_worker_hook(TileWorkerData *const tile_data,
 static int compare_tile_buffers(const void *a, const void *b) {
   const TileBuffer *const buf1 = (const TileBuffer*)a;
   const TileBuffer *const buf2 = (const TileBuffer*)b;
-  return (int)(buf2->size - buf1->size);
+  return (int)(buf2->size) - (int)(buf1->size);
 }
 
 static const uint8_t *decode_tiles_mt(VP9Decoder *pbi,
